+++ cmd/hidden_lake/helpers/loader/internal/handler/transfer_test.go
@@ -139,8 +139,6 @@ func TestHandleTransferAPI(t *testing.T) {
 				hls_settings.CNetworkMask,
 				encMsg.ToBytes(),
 			),
-			1,
-			0,
 		)
 		err = hltClient1.PutMessage(context.Background(), netMsg)
 		if err != nil {