+++ cmd/hidden_lake/helpers/traffic/internal/handler/message_test.go
@@ -36,8 +36,6 @@ func TestHandleMessageAPI(t *testing.T) {
 	netMsg := net_message.NewMessage(
 		testNetworkMessageSettings(),
 		payload.NewPayload64(hls_settings.CNetworkMask, msg.ToBytes()),
-		1,
-		0,
 	)
 	if err := hltClient.PutMessage(context.Background(), netMsg); err != nil {
 		t.Error(err)