+++ examples/traffic_actions/keeper/client_hlt/main.go
@@ -87,8 +87,6 @@ func main() {
 		netMsg := net_message.NewMessage(
 			netSett,
 			payload.NewPayload64(hls_settings.CNetworkMask, msg.ToBytes()),
-			1,
-			0,
 		)
 
 		if err := hltClient.PutMessage(ctx, netMsg); err != nil {