+++ pkg/network/examples/ping-pong/main.go
@@ -60,8 +60,6 @@ func main() {
 			serviceHeader,
 			[]byte("0"),
 		),
-		1,
-		0,
 	)
 	service2.BroadcastMessage(ctx, msg)
 
@@ -94,8 +92,6 @@ func handler(serviceName string) network.IHandlerF {
 					serviceHeader,
 					[]byte(fmt.Sprintf("%d", num+1)),
 				),
-				1,
-				0,
 			),
 		)