+++ pkg/network/message/settings.go
@@ -1,23 +1,27 @@
 package message
 
 var (
-	_ ISettings = &sSettings{}
+	_ IConstructSettings = &sSettings{}
 )
 
 type SSettings sSettings
 type sSettings struct {
-	FWorkSizeBits uint64
-	FNetworkKey   string
+	FWorkSizeBits       uint64
+	FNetworkKey         string
+	FParallel           uint64
+	FLimitVoidSizeBytes uint64
 }
 
-func NewSettings(pSett *SSettings) ISettings {
+func NewSettings(pSett *SSettings) IConstructSettings {
 	return (&sSettings{
-		FWorkSizeBits: pSett.FWorkSizeBits,
-		FNetworkKey:   pSett.FNetworkKey,
+		FWorkSizeBits:       pSett.FWorkSizeBits,
+		FNetworkKey:         pSett.FNetworkKey,
+		FParallel:           pSett.FParallel,
+		FLimitVoidSizeBytes: pSett.FLimitVoidSizeBytes,
 	}).mustNotNull()
 }
 
-func (p *sSettings) mustNotNull() ISettings {
+func (p *sSettings) mustNotNull() IConstructSettings {
 	return p
 }
 
@@ -27,4 +31,12 @@ func (p *sSettings) GetWorkSizeBits() uint64 {
 
 func (p *sSettings) GetNetworkKey() string {
 	return p.FNetworkKey
+}
+
+func (p *sSettings) GetParallel() uint64 {
+	return p.FParallel
+}
+
+func (p *sSettings) GetLimitVoidSizeBytes() uint64 {
+	return p.FLimitVoidSizeBytes
 }